requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
flask-cors==4.0.0
python-whois==0.9.4
dnspython==2.6.1
